[REFACTOR] move more stuff out of game.hpp
remaining entities stuff go to entities.hpp and mastermode stuff to network/legacy/administration.hpp
