+++ inexor/fpsgame/ai.cpp
@@ -1,4 +1,6 @@
 #include "inexor/fpsgame/game.hpp"
+#include "inexor/fpsgame/entities.hpp"
+
 
 extern SharedVar<int> fog;
